Server/requirements.txt
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
click==8.0.1
colorama==0.4.4
distlib==0.3.3
dnspython==2.1.0
elastic-transport==8.1.2
elasticsearch==7.14.1
email-validator==1.1.3
filelock==3.3.2
Flask==2.0.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.4.0
pymongo==3.12.1
python-dotenv==0.19.1
six==1.16.0
toml==0.10.2
urllib3==1.26.6
virtualenv-clone==0.5.7
Werkzeug==2.0.1
WTForms==2.3.3
pywebio
rapidfuzz
mhcnames
requests
spacy
pandas
beautifulsoup4
Bio